double sum

+++ src/main/java/by/epam/taskThree/service/ConvertService.java
@@ -3,7 +3,7 @@ package by.epam.taskThree.service;
 import java.util.List;
 
 public class ConvertService {
-    public int[] arrayListToArray(List<Integer> integerList) {
-        return integerList.stream().mapToInt(i -> i).toArray();
+    public double[] arrayListToArray(List<Double> integerList) {
+        return integerList.stream().mapToDouble(i -> i).toArray();
     }
 }
